pyproject.toml
[tool.poetry]
name = "democrance-api-test"
version = "0.1.0"
description = "A simple API that has a few REST endpoints."
authors = ["Saad Aleem <aleemsaadullah@gmail.com>"]
readme = "README.md"
packages = [{include = "democranceapi"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
pytest = "^7.2.0"
psycopg2 = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.25.1"
django-simple-history = "^3.2.0"
django-filter = "^22.1"
model_bakery = "^1.10.1"
Faker = "^16.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
